+++ tidal_api/routes/tracks.py
@@ -1,0 +1,93 @@
+"""Track and recommendation route implementation logic."""
+import concurrent.futures
+from tidal_api.browser_session import BrowserSession
+from tidal_api.utils import format_track_data, bound_limit
+
+
+def get_user_tracks(session: BrowserSession, limit: int = 10) -> dict:
+    """Implementation logic for getting user's favorite tracks."""
+    try:
+        favorites = session.user.favorites
+        limit = bound_limit(limit)
+
+        tracks = favorites.tracks(limit=limit, order="DATE", order_direction="DESC")
+        track_list = [format_track_data(track) for track in tracks]
+
+        return {"tracks": track_list}, 200
+    except Exception as e:
+        return {"error": f"Error fetching tracks: {str(e)}"}, 500
+
+
+def get_single_track_recommendations(
+    session: BrowserSession,
+    track_id: str,
+    limit: int = 10
+) -> dict:
+    """Implementation logic for getting recommendations for a single track."""
+    try:
+        limit = bound_limit(limit)
+
+        track = session.track(track_id)
+        if not track:
+            return {"error": f"Track with ID {track_id} not found"}, 404
+
+        recommendations = track.get_track_radio(limit=limit)
+        track_list = [format_track_data(track) for track in recommendations]
+
+        return {"recommendations": track_list}, 200
+    except Exception as e:
+        return {"error": f"Error fetching recommendations: {str(e)}"}, 500
+
+
+def get_batch_track_recommendations(
+    session: BrowserSession,
+    track_ids: list,
+    limit_per_track: int = 20,
+    remove_duplicates: bool = True
+) -> dict:
+    """Implementation logic for getting batch recommendations."""
+    try:
+        if not isinstance(track_ids, list):
+            return {"error": "track_ids must be a list"}, 400
+
+        limit_per_track = bound_limit(limit_per_track)
+
+        def get_track_recommendations(track_id):
+            """Function to get recommendations for a single track"""
+            try:
+                track = session.track(track_id)
+                recommendations = track.get_track_radio(limit=limit_per_track)
+                formatted_recommendations = [
+                    format_track_data(rec, source_track_id=track_id)
+                    for rec in recommendations
+                ]
+                return formatted_recommendations
+            except Exception as e:
+                print(f"Error getting recommendations for track {track_id}: {str(e)}")
+                return []
+
+        all_recommendations = []
+        seen_track_ids = set()
+
+        # Use ThreadPoolExecutor to process tracks concurrently
+        with concurrent.futures.ThreadPoolExecutor(max_workers=len(track_ids)) as executor:
+            future_to_track_id = {
+                executor.submit(get_track_recommendations, track_id): track_id
+                for track_id in track_ids
+            }
+
+            for future in concurrent.futures.as_completed(future_to_track_id):
+                track_recommendations = future.result()
+
+                for track_data in track_recommendations:
+                    track_id = track_data.get('id')
+
+                    if remove_duplicates and track_id in seen_track_ids:
+                        continue
+
+                    all_recommendations.append(track_data)
+                    seen_track_ids.add(track_id)
+
+        return {"recommendations": all_recommendations}, 200
+    except Exception as e:
+        return {"error": f"Error fetching batch recommendations: {str(e)}"}, 500
